reverted mercurius version bump, moved @defer tests to their own file
