+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "mercurius",
-  "version": "11.2.0",
+  "version": "11.1.0",
   "description": "Fastify GraphQL adapter with gateway and subscription support",
   "main": "index.js",
   "types": "index.d.ts",
